refactor: 完成 REVIEW 修复收敛
按 REVIEW.md remediation 计划收敛入口、平台、控制面、dataplane、products、Web/Admin/WebUI 的问题与测试覆盖。

验证: rtk go test ./...; rtk go test -race ./app/products/web/...; rtk git diff --check

+++ app/control/account/backends/redis_adapter.go
@@ -74,7 +74,7 @@ func (s *goRedisAccountStore) HSet(ctx context.Context, key string, mapping map[
 	return s.client.HSet(ctx, key, redisHashArgs(mapping)...).Err()
 }
 
-func (s *goRedisAccountStore) ZAdd(ctx context.Context, key string, members map[string]int) error {
+func (s *goRedisAccountStore) ZAdd(ctx context.Context, key string, members map[string]int64) error {
 	values := make([]redis.Z, 0, len(members))
 	for member, score := range members {
 		values = append(values, redis.Z{Member: member, Score: float64(score)})
